// src/app/flows/cases/cases-create-casefile/components/cases-create-casefile-bank-details/interfaces/cases-create-casefile-bank-details-field-names.interface.ts
export interface ICasesCreateCasefileBankDetailsFieldNames {
  bankType: string;
  ukNameOnAccount: string;
  ukSortCode: string;
  ukAccountNumber: string;
  ukPaymentReference: string;
  nonUkNameOnAccount: string;
  nonUkAccountNumber: string;
  nonUkPaymentReference: string;
  nonUkBicSwiftCode: string;
  nonUkIban: string;
  nonUkBankName: string;
  nonUkBranchSortCode: string;
}